+++ changelog.md
@@ -1,2 +1,2 @@
 ### Scarecrows' Territory 1.1.9
-- Account for changes in Forge 45.0.23
+- Initial release of Scarecrows' Territory for Minecraft 1.20